backend/seguridad_servicio/gestor_usuarios/requirements.txt
blinker==1.9.0
click==8.1.8
coverage==7.7.0
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
PyJWT==2.6.0
Flask
Flask-SQLAlchemy
psycopg2-binary
flask-cors==5.0.1
pycryptodome